# Change auto-deletion throttling constants to better scale with server size