Note: this is updated using sparse data from medium-sized servers. I would like to see data from larger servers like mastodon.social to better fine-tune those constants.

- `MAX_BUDGET`: increase from 50 per minute to 150 per minute, as 50 per minute might be too low to keep up with mastodon.social's 56~96 new posts per minute (though the number of toots to be auto-deleted is probably far lower)
- `PER_THREAD_BUDGET`: increase from 5 to 6 per minute, this is more to give more leeway to small servers
- `MAX_DEFAULT_LATENCY`: increase from 5s to 8s, to give more chances to automatic deletions to be queued, without affecting server load too much
- `MAX_PUSH_LATENCY`: increase from 10s to 15s, to give more chances to automatic deletions to be queued, without affecting server load too much
- `MAX_DEFAULT_SIZE`: increase from 2 to 200 items, as larger servers will have larger queues (while still quickly chewing through them), a limit of 2 items is not reasonable, as a server with any medium activity will often have more than that
- `MAX_PUSH_SIZE`: increase from 5 to 500 items, as larger servers will have larger queues (while still quickly chewing through them), a limit of 5 items is not reasonable, as a server with any medium activity will often have more than that
- `MAX_RETRY_SIZE`: I bumped it, but honestly I think I need more data